+++ src/Raven.CodeAnalysis/CodeGen/MethodBodyGenerator.cs
@@ -2738,6 +2738,8 @@ internal class MethodBodyGenerator
         {
             BoundReturnStatement => true,
             BoundThrowStatement => true,
+            BoundExpressionStatement { Expression: BoundInvocationExpression invocation }
+                when BoundNodeFacts.IsTerminalRuntimeFault(invocation.Method) => true,
             BoundExpressionStatement { Expression: BoundReturnExpression or BoundThrowExpression } => true,
             BoundBlockStatement block when block.Statements.Any() => IsTerminatingStatement(block.Statements.Last()),
             BoundIfStatement { ElseNode: not null } ifStatement =>